+++ Examples/beacons/Examples/ESTBeaconDetailsDemoVC.h
@@ -1,0 +1,14 @@
+//
+//  ESTBeaconDetailsDemoVC.h
+//
+//  Copyright (c) 2015 Estimote. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+#import "ESTBeacon.h"
+
+@interface ESTBeaconDetailsDemoVC : UIViewController
+
+- (id)initWithBeacon:(ESTBeacon*)beacon;
+
+@end